ath10k: OTP debugging for CT firmware.
OTP firmware image does not have any logging support or
other debugging on normal hardware.

With the CT firmware, the OTP may choose to return a pointer
to it's memory location that contains debug information.

This patch enables the driver to read this memory after
the OTP has exited to aid debugging.

Signed-off-by: Ben Greear <greearb@candelatech.com>