ath10k: add SURVEY_INFO_IN_USE for current channel on survey
When user space requests survey info, it is useful to know which of the survey
data refers to the channel that is currently actively being used.  One of the
use cases is getting the current channel noise for status output.  Without this
flag you would have to look up the channel separately and then compare it
against the frequency in the survey output in user space.

Signed-off-by: Felix Fietkau <nbd at openwrt.org>
Signed-off-by: Kalle Valo <kvalo at qca.qualcomm.com>