ath10k:  Allow setting bcast, mcast, mgt fixed tx rates.
The default behaviour when setting a fixed transmit rate in ath10k
is to only set the 'ucast' rate.  This does not affect beacons,
multicast, or broadcast traffic.

In order to allow a back-door way to set these other types of
traffic as well, provide a debugfs file that can set the
traffic type.  Then, use can use the normal 'iw' commands to
set the rate to the desired value.

Upstream firmware *should* support all of this, but at least
10.1 does NOT have support for the WMI_10X_VDEV_PARAM_MGMT_RATE
parameter.  I added this to the CT firmware so it works there.

Possibly other upstream firmware have fixed this as well.

Signed-off-by: Ben Greear <greearb@candelatech.com>