ath10k:  Allow latching max tx-power (CT only)
This allows a one-time set of the maximum tx-power
(per module load, so compile the module into the kernel
if you don't want to allow this to be reset w/out a reboot).

The value is propagated to the firmware, including on
firmware resets.

The purpose is to have a tool set the max power early on
boot, and then any 'iw' or other settings that try to set
a higher tx power will simply be ignored by the firmware.

Requires beta-15 CT firmware built on or after Oct 15, 2015
to actually function as described.

Example that sets max txpower to 16db:

echo 0x600000020 > /debug/ieee80211/wiphy0/ath10k/ct_special

The 0x6 is the identifier, the 0x20 component is the power
in 0.5 db units.

Signed-off-by: Ben Greear <greearb@candelatech.com>